requirements.txt
streamlit==1.32.0
google-generativeai==0.3.2
python-dotenv==1.0.0
crewai==0.11.0
langchain==0.1.9 